added intense shake on prize win, also transferred to another component to lessen bloat
